+++ compose.yml
@@ -45,7 +45,6 @@ services:
       - node_modules_ghost_domain-events:/home/ghost/ghost/domain-events/node_modules:delegated
       - node_modules_ghost_donations:/home/ghost/ghost/donations/node_modules:delegated
       - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
-      - node_modules_ghost_email-analytics-service:/home/ghost/ghost/email-analytics-service/node_modules:delegated
       - node_modules_ghost_email-content-generator:/home/ghost/ghost/email-content-generator/node_modules:delegated
       - node_modules_ghost_email-events:/home/ghost/ghost/email-events/node_modules:delegated
       - node_modules_ghost_email-service:/home/ghost/ghost/email-service/node_modules:delegated
@@ -192,7 +191,6 @@ volumes:
   node_modules_ghost_domain-events: {}
   node_modules_ghost_donations: {}
   node_modules_ghost_email-addresses: {}
-  node_modules_ghost_email-analytics-service: {}
   node_modules_ghost_email-content-generator: {}
   node_modules_ghost_email-events: {}
   node_modules_ghost_email-service: {}
