+++ ghost/core/core/server/services/email-analytics/EventProcessingResult.js
