+++ ghost/core/test/unit/server/services/email-analytics/EmailAnalyticsProviderMailgun.test.js
@@ -1,8 +1,6 @@
 const sinon = require('sinon');
 
-const {EventProcessingResult} = require('@tryghost/email-analytics-service');
-
-// module under test
+const EventProcessingResult = require('../../../../../core/server/services/email-analytics/EventProcessingResult');
 const EmailAnalyticsProviderMailgun = require('../../../../../core/server/services/email-analytics/EmailAnalyticsProviderMailgun');
 
 const SAMPLE_EVENTS = [
