--- ghost/email-analytics-service/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    EmailAnalyticsService: require('./lib/EmailAnalyticsService'),
-    EventProcessingResult: require('./lib/EventProcessingResult')
-};
